set anchorUSD code and address for escrow
